Small fixes for alpine build

+++ libs/oef-base/include/oef-base/comms/CharArrayBuffer.hpp
@@ -88,7 +88,7 @@ public:
     } buffer;
 
     // buffer.i = i;
-    buffer.i = static_cast<int32_t>(htonl(i));
+    buffer.i = static_cast<int32_t>(htonl(static_cast<uint32_t>(i)));
     oflow(buffer.c[0]);
     oflow(buffer.c[1]);
     oflow(buffer.c[2]);
@@ -109,7 +109,7 @@ public:
     buffer.c[1] = (uint8_t)uflow();
     buffer.c[2] = (uint8_t)uflow();
     buffer.c[3] = (uint8_t)uflow();
-    i           = static_cast<int32_t>(ntohl(buffer.i));
+    i           = static_cast<int32_t>(ntohl(static_cast<uint32_t>(buffer.i)));
     // i = buffer.i;
 
     return *this;
